better loading and configuration the logger module
