+++ lib/JumpScale/core/logging/LoggerFactory.py
@@ -1,5 +1,8 @@
 from JumpScale import j
 
+from JSLogger import JSLogger
+from Filter import ModuleFilter
+
 import logging
 import logging.handlers
 from colorlog import ColoredFormatter
@@ -34,62 +37,13 @@ class LoggerFactory:
         self.PRODUCTION = PRODUCTION
         self.DEV = DEV
 
-        # TODO read mode from config.
-        # Currently the init of jumpscale trigger an infinite loop if I try
-        # to read system hrd here.
-        if True:
-            self.set_mode(self.DEV)
-
-    def _enable_production_mode(self):
-        self._logger.handlers = []
-        self._logger.addHandler(logging.NullHandler())
-        self._logger.propagate = True
-
-    def _enable_dev_mode(self):
-        logging.setLoggerClass(JSLogger)
         self._logger = logging.getLogger(self.root_logger_name)
-        self._logger.setLevel(logging.DEBUG)
-        self._logger.propagate = False
-        logging.lastResort = None
+        self._logger.addHandler(logging.NullHandler())
 
-        self.handlers = {
-            "console": self.__consoleHandler(),
-            "file": self.__fileRotateHandler(),
-        }
-        for h in self.handlers.values():
-            self._logger.addHandler(h)
-
-    def __fileRotateHandler(self, name='jumpscale'):
-        filename = "/optvar/log/%s.log" % name  # TODO can't use j.dirs here before j.dirs is loaded.
-        formatter = logging.Formatter(FILE_FORMAT)
-        fh = logging.handlers.TimedRotatingFileHandler(filename, when='D', interval=1, backupCount=7, encoding=None, delay=False, utc=False, atTime=None)
-        fh.setLevel(logging.DEBUG)
-        fh.setFormatter(formatter)
-        return fh
-
-    def __consoleHandler(self):
-        formatter = ColoredFormatter(
-            fmt=CONSOLE_FORMAT,
-            datefmt="%a%d %H:%M",
-            reset=True,
-            log_colors={
-                'DEBUG':    'cyan',
-                'INFO':     'green',
-                'WARNING':  'yellow',
-                'ERROR':    'red',
-                'CRITICAL': 'red,bg_white',
-            },
-            secondary_log_colors={},
-            style='%'
-        )
-        ch = logging.StreamHandler()
-        ch.setLevel(logging.INFO)
-        ch.setFormatter(formatter)
-        return ch
-
-    def __redisHandler(self, redis_client=None):
-            if redis_client is None:
-                self.redis_client = j.core.db
+    def init(self, mode, level, filter=[]):
+        self.set_mode(mode.upper())
+        self.set_level(level)
+        self._logger.addFilter(ModuleFilter(filter))
 
     def get(self, name=None, enable_only_me=False):
         """
@@ -124,70 +78,64 @@ class LoggerFactory:
         elif mode == self.DEV:
             self._enable_dev_mode()
 
+    def set_level(self, level):
+        self._logger.setLevel(level)
+
     def log(self, msg=None, level=None, category=None):
         self._logger.log(level, msg)
 
+    def _enable_production_mode(self):
+        self._logger.handlers = []
+        self._logger.addHandler(logging.NullHandler())
+        self._logger.propagate = True
 
-class JSLogger(logging.Logger):
+    def _enable_dev_mode(self):
+        logging.setLoggerClass(JSLogger)
+        self._logger.setLevel(logging.DEBUG)
+        self._logger.propagate = False
+        logging.lastResort = None
+        self.handlers = {
+            "console": self.__consoleHandler(),
+            "file": self.__fileRotateHandler(),
+        }
+        for h in self.handlers.values():
+            self._logger.addHandler(h)
 
-    def __init__(self, name):
-        super(JSLogger, self).__init__(name)
-        self.custom_filters = {}
-        self.__only_me = False
+    def __fileRotateHandler(self, name='jumpscale'):
+        filename = "/optvar/log/%s.log" % name  # TODO can't use j.dirs here before j.dirs is loaded.
+        formatter = logging.Formatter(FILE_FORMAT)
+        fh = logging.handlers.TimedRotatingFileHandler(filename, when='D', interval=1, backupCount=7, encoding=None, delay=False, utc=False, atTime=None)
+        fh.setLevel(logging.DEBUG)
+        fh.setFormatter(formatter)
+        return fh
 
-    def error(self, msg, *args, **kwargs):
-        """
-        Log 'msg % args' with severity 'ERROR'.
+    def __consoleHandler(self):
+        formatter = ColoredFormatter(
+            fmt=CONSOLE_FORMAT,
+            datefmt="%a%d %H:%M",
+            reset=True,
+            log_colors={
+                'DEBUG':    'cyan',
+                'INFO':     'green',
+                'WARNING':  'yellow',
+                'ERROR':    'red',
+                'CRITICAL': 'red,bg_white',
+            },
+            secondary_log_colors={},
+            style='%'
+        )
+        ch = logging.StreamHandler()
+        ch.setLevel(logging.INFO)
+        ch.setFormatter(formatter)
+        logging_cfg = j.data.hrd.get('/optvar/hrd/system/logging.hrd')
+        modules = logging_cfg.get('filter')
+        ch.addFilter(ModuleFilter(modules))
+        return ch
 
-        To pass exception information, use the keyword argument exc_info with
-        a true value, e.g.
+    # def __filter_module(self, modules):
+    #     def filter(record):
+    #         if record.name:
 
-        logger.error("Houston, we have a %s", "major problem", exc_info=1)
-
-        """
-        if self.isEnabledFor(logging.ERROR):
-            eco = j.errorconditionhandler.getErrorConditionObject(
-                ddict={}, msg=msg, msgpub=msg, category=self.name,
-                level=logging.ERROR, type=logging.getLevelName(logging.ERROR),
-                tb=None, tags='')
-            j.errorconditionhandler._send2Redis(eco)
-
-            self._log(logging.ERROR, msg, args, **kwargs)
-
-    def critical(self, msg, *args, **kwargs):
-        """
-        Log 'msg % args' with severity 'CRITICAL'.
-
-        To pass exception information, use the keyword argument exc_info with
-        a true value, e.g.
-
-        logger.critical("Houston, we have a %s", "major disaster", exc_info=1)
-        """
-        if self.isEnabledFor(CRITICAL):
-            eco = j.errorconditionhandler.getErrorConditionObject(
-                ddict={}, msg=msg, msgpub=msg, category=self.name,
-                level=logging.CRITICAL, type=logging.getLevelName(logging.CRITICAL),
-                tb=None, tags='')
-            j.errorconditionhandler._send2Redis(eco)
-
-            self._log(logging.CRITICAL, msg, args, **kwargs)
-
-
-    def enable_only_me(self):
-        """
-        Enable filtering. Output only log from this logger and its children.
-        Logs from other modules are masked
-        """
-        if not self.__only_me and 'console' in j.logger.handlers:
-            only_me_filter = logging.Filter(self.name)
-            j.logger.handlers['console'].addFilter(only_me_filter)
-            self.custom_filters["only_me"] = only_me_filter
-            self.__only_me = True
-
-    def disable_only_me(self):
-        """
-        Disable filtering on only this logger
-        """
-        if self.__only_me and 'console' in j.logger.handlers:
-            j.logger.handlers['console'].removeFilter(self.custom_filters['only_me'])
-            self.__only_me = False
+    def __redisHandler(self, redis_client=None):
+            if redis_client is None:
+                self.redis_client = j.core.db
